requirements.txt
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
chromadb
pypdf
openai
sentence-transformers
umap-learn
matplotlib
pandas
langchain
langchain-community
streamlit
python-dotenv
tqdm
